went back to relative file path with .. & reorganized stuff under htr and downloaded + set up ctcwordbeamsearch
Relative file path with .. works because we are expected to use command line to run this, where the directory will be set to the main.py file. Need to find a way to have run and output work without typing into command line or receiving output in command line. Have to also work on making word beam search the default decoder after setting up the IAM model.